feat: add validation for contactId
