+++ routes/api/contacts.js
@@ -1,23 +1,14 @@
 const express = require("express");
 const router = express.Router();
 const {
-  listContacts,
-  getContactById,
   addContact,
+  getContactById,
+  listContacts,
   removeContact,
   updateContact,
 } = require("../../models/contacts");
-const Joi = require("joi");
 
-const contactSchema = Joi.object({
-  name: Joi.string().trim().min(1).max(50).required(),
-  email: Joi.string().email().required(),
-  phone: Joi.string()
-    .pattern(/^[0-9]+$/)
-    .min(9)
-    .max(15)
-    .required(),
-});
+const { contactSchema, contactIdSchema } = require("./validation");
 
 router.get("/", async (req, res, next) => {
   try {
@@ -56,6 +47,11 @@ router.post("/", async (req, res, next) => {
 
 router.delete("/:contactId", async (req, res, next) => {
   try {
+    const { error } = contactIdSchema.validate(req.params.contactId);
+    if (error) {
+      return res.status(400).json({ message: error.details[0].message });
+    }
+
     const removedContact = await removeContact(req.params.contactId);
     if (!removedContact) {
       return res.status(404).json({ message: "Contact not found" });
@@ -68,9 +64,14 @@ router.delete("/:contactId", async (req, res, next) => {
 
 router.put("/:contactId", async (req, res, next) => {
   try {
-    const { error } = contactSchema.validate(req.body);
-    if (error) {
-      return res.status(400).json({ message: error.details[0].message });
+    const { error: idError } = contactIdSchema.validate(req.params.contactId);
+    if (idError) {
+      return res.status(400).json({ message: idError.details[0].message });
+    }
+
+    const { error: bodyError } = contactSchema.validate(req.body);
+    if (bodyError) {
+      return res.status(400).json({ message: bodyError.details[0].message });
     }
 
     const updatedContact = await updateContact(req.params.contactId, req.body);
